requirements.txt
Flask
Flask-PyMongo
pandas
plotly
tqdm
gunicorn
python-dotenv
mlxtend
seaborn